+++ R/utilities.R
@@ -21,21 +21,24 @@ type_by_datauri <- function(x){
   types <- get_types(x)
   r_types <- vapply(get_types(x), r_class, character(length(1)))
   df <- data.frame(value = get_values(x), class = r_types)
-  apply(df, 1, function(x) as(x[1], x[2]))
+  apply(df, 1, function(x) as(utf8me(x[1]), x[2]))
 }
 
+#' @importFrom stringi stri_unescape_unicode
+utf8me <- function(x){ 
+  removed_quotes <- gsub('\"', '', x)
+  stringi::stri_unescape_unicode(removed_quotes)
+}
 
 rectangularize_query_results <- function(out){
   vars <- unique(names(out))
-  X <- lapply(vars, function(v) 
-    ## Strip ^^TYPE typing
-    #gsub('\"(([^\\^])+)\"\\^*.*', 
-    #     "\\1", 
-     type_by_datauri(as.character(out[names(out) == v ])))
+  X <- lapply(vars, function(v){ 
+     contents <- as.character(out[names(out) == v ])
+     type_by_datauri(contents)
+  })
   names(X) <- vars
   as.data.frame(X, stringsAsFactors=FALSE)
 }
-
 
 
 
@@ -68,9 +71,10 @@ rdf_mimetypes <- c("nquads" = "text/x-nquads",
                    "trig" = "application/x-trig",
                    "turtle" = "text/turtle")
 
-# trig not working right now, not clear why
-# Consider adding/testing: 
-# - n3 (text/n3)
+## My redland version does not find support for these, probably optional 
+## additional dependency needed when compiling redland
+# - trig  (application/x-trig)
+# - n3 (text/n3) 
 # - rdfa (application/xhtml+xml, or text/html)
 # - rss (application/rss+xml or text/rss)